Update exception handling

+++ src/main/java/com/vikash/restraunt/controller/RestrauntRestController.java
@@ -4,6 +4,7 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 
 import org.slf4j.Logger;
@@ -15,10 +16,12 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.vikash.restraunt.entities.CuisineStyle;
 import com.vikash.restraunt.entities.Restraunt;
+import com.vikash.restraunt.exception.handling.RestrauntNotFoundException;
 import com.vikash.restraunt.repos.RestrauntRepository;
 
 @RestController
@@ -30,6 +33,7 @@ public class RestrauntRestController {
 	RestrauntRepository restrauntRepository;
 
 	@PostMapping("/create")
+	@ResponseBody
 	public Restraunt createRestraunt(@RequestBody Restraunt restraunt) {
 		
 		LOGGER.info("Creating restraunt of the following fields " + restraunt);
@@ -39,7 +43,14 @@ public class RestrauntRestController {
 	}
 	
 	@DeleteMapping("/delete")
+	@ResponseBody
 	public String deleteRestraunt(@RequestParam("id") int id) {
+		
+		Optional<Restraunt> restraunt = restrauntRepository.findById(id);
+		if (!restraunt.isPresent()) {
+			LOGGER.error("Id not found id: " + id);
+			throw new RestrauntNotFoundException("Id not found id: " + id);
+		}
 		
 		LOGGER.info("Delete restraunt of id = " + id);
 		restrauntRepository.deleteById(id);
@@ -48,6 +59,7 @@ public class RestrauntRestController {
 	}
 
 	@GetMapping("/getAll")
+	@ResponseBody
 	public List<Restraunt> getAllRestraunt() {
 		LOGGER.info("Get all the restraunt");
 		List<Restraunt> restraunts = restrauntRepository.findAll();
@@ -55,22 +67,33 @@ public class RestrauntRestController {
 	}
 
 	@GetMapping("/name/{name}")
-	public Restraunt searchByName(@PathVariable("name") String name) {
+	@ResponseBody
+	public Optional<Restraunt> searchByName(@PathVariable("name") String name) {
 		
 		LOGGER.info("Search restraunt by restraunt name " + name);
-		Restraunt restraunt = restrauntRepository.findByName(name);
+		Optional<Restraunt> restraunt = restrauntRepository.findByName(name);
+		if (!restraunt.isPresent()) {
+			LOGGER.error("Name not found name: " + name);
+			throw new RestrauntNotFoundException("Name not found name: " + name);
+		}
 		return restraunt;
 	}
 
 	@GetMapping("/city/{city}")
+	@ResponseBody
 	public List<Restraunt> searchByCity(@PathVariable("city") String city) {
 
 		LOGGER.info("Search restraunt by restraunt city " + city);
 		List<Restraunt> restraunts = restrauntRepository.findByCity(city);
+		if (restraunts.size() == 0) {
+			LOGGER.info("No data restraunt present in " + city);
+			throw new RestrauntNotFoundException("No data restraunt present in " + city);
+		}
 		return restraunts;
 	}
 
 	@GetMapping("/sort/ranking")
+	@ResponseBody
 	public List<Restraunt> sortByRanking() {
 		
 		LOGGER.info("Sort restraunt by restraunt ranking");
@@ -88,6 +111,7 @@ public class RestrauntRestController {
 	}
 
 	@GetMapping("/sort/rating")
+	@ResponseBody
 	public List<Restraunt> sortByRating() {
 		
 		LOGGER.info("Sort restraunt by restraunt rating");
@@ -105,6 +129,7 @@ public class RestrauntRestController {
 	}
 
 	@PostMapping("/filter/style")
+	@ResponseBody
 	public Set<Restraunt> filterByStyle(@RequestBody CuisineStyle style) {
 		
 		LOGGER.info("Filter Cuisine Style of restraunt " + style);
@@ -122,6 +147,12 @@ public class RestrauntRestController {
 				
 			}
 			
+		}
+		
+		if (results.size() == 0) {
+			
+			LOGGER.info("No data restraunt serve " + style + " style");
+			throw new RestrauntNotFoundException("No data restraunt serve " + style + " style");
 		}
 		
 		return results;
